requirements.txt
click
python-dotenv
wolframalpha
PyGithub
